Required fields added in Order, JDBC template refactored
